+++ pkg/gen/input/workflows/workflows.go
@@ -49,16 +49,8 @@ func (w *Workflows) CreateRelease() input.Input {
 	return file.NewCreateReleaseInput(w.params)
 }
 
-func (w *Workflows) CreateReleaseDeletion() input.Input {
-	return file.NewCreateReleaseDeletionInput(w.params)
-}
-
 func (w *Workflows) CreateReleasePR() input.Input {
 	return file.NewCreateReleasePRInput(w.params)
-}
-
-func (w *Workflows) CreateReleasePRDeletion() input.Input {
-	return file.NewCreateReleasePRDeletionInput(w.params)
 }
 
 func (w *Workflows) DispatchUpdateChartEvents(targetRepo string) input.Input {
@@ -105,26 +97,10 @@ func (w *Workflows) ValidateChangelog() input.Input {
 	return file.NewValidateChangelogInput(w.params)
 }
 
-func (w *Workflows) ValidateChangelogDeletion() input.Input {
-	return file.NewValidateChangelogDeletionInput(w.params)
-}
-
 func (w *Workflows) AnalyzeGithubActions() input.Input {
 	return file.NewAnalyzeGithubActionsInput(w.params)
 }
 
 func (w *Workflows) ZizmorBaseYml() input.Input {
 	return file.NewZizmorBaseInput(w.params)
-}
-
-func (w *Workflows) ReleasePlease(autoMergeLevel string) input.Input {
-	return file.NewReleasePleaseInput(w.params, autoMergeLevel)
-}
-
-func (w *Workflows) ReleasePleaseConfig(changelogStyle string, hasProjectGo bool) input.Input {
-	return file.NewReleasePleaseConfigInput(changelogStyle, hasProjectGo)
-}
-
-func (w *Workflows) ReleasePleaseManifest() input.Input {
-	return file.NewReleasePleaseManifestInput()
 }